pyproject.toml
[tool.poetry]
name = "mongodblab"
version = "0.1.0"
description = ""
authors = ["Christhoval Barba <christhoval@dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pymongo = "^4.7.1"
python-dateutil = "^2.9.0.post0"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
